Changed Product.Price to decimal type

+++ CrudApi/Models/Product.cs
@@ -1,3 +1,3 @@
 namespace CrudApi.Models;
 
-public record Product(int Id, string Name, int Price);
+public record Product(int Id, string Name, decimal Price);
